Completed all data service methods, separated Credentials and User data service

+++ src/com/jingle/business/UserBusinessService.java
@@ -5,7 +5,9 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.jingle.data.CredentialsDataInterface;
 import com.jingle.data.UserDataInterface;
+import com.jingle.model.Credentials;
 import com.jingle.model.User;
 
 /**
@@ -18,61 +20,114 @@ public class UserBusinessService implements UserBusinessInterface {
 
 	UserDataInterface userDataService;
 
+	CredentialsDataInterface credentialsDataService;
+
 	@Autowired
 	public void setUserDataService(UserDataInterface userDataService) {
 		this.userDataService = userDataService;
 	}
 
+	@Autowired
+	public void setCredentialsDataService(CredentialsDataInterface credentialsDataService) {
+		this.credentialsDataService = credentialsDataService;
+	}
+
 	/**
-	 * Takes in a user. 
-	 * Returns the userDataService create method with user as the parameter.
+	 * Take in a user. 
+	 * Create credentials for user with credentialsDataService. 
+	 * If not created, return -1.  
+	 * Return the userDataService create method using user. 
 	 * 
 	 * @param user	user to register
 	 * @return int	result
 	 */
 	public int registerUser(User user) {
+		int credentials_id = credentialsDataService.create(user.getCredentials());
+
+		if (credentials_id < 0) {
+			return -1;
+		}
+
+		user.setCredentials_id(credentials_id);
+
 		return userDataService.create(user);
 	}
 
 	/**
-	 * Takes in a user. 
-	 * Returns the userDataService readByCredentials method with user as the parameter. 
+	 * Take in a user. 
+	 * Select credentials for user with credentialsDataService. 
+	 * If not selected, return null. 
+	 * Set user's credentials. 
+	 * Return the userDataService readByCredentialsId method using user. 
 	 * 
-	 * @param user 	user to login
-	 * @return User user that logged in
+	 * @param user	user to login
+	 * @return int	result
 	 */
 	public User loginUser(User user) {
-		return userDataService.readByCredentials(user);
+
+		Credentials credentials = credentialsDataService.readByUsernamePassword(user.getCredentials());
+
+		if (credentials == null) {
+			return null;
+		}
+
+		user.setCredentials(credentials);
+		user.setCredentials_id(credentials.getId());
+
+		return userDataService.readByCredentialsId(user);
 	}
 
 	/**
-	 * Returns the userDataService readAll method.
+	 * Select all users with userDataService readAll. 
+	 * For each user, set its credentials with credentialsDataService read. 
+	 * Return all users. 
 	 * 
 	 * @return List<User>	list of all users
 	 */
 	public List<User> getAllUsers() {
-		return userDataService.readAll();
+
+		List<User> users = userDataService.readAll();
+		
+		for(User user : users) {
+			user.setCredentials(credentialsDataService.read(new Credentials(user.getCredentials_id(), "", "")));
+		}
+		
+		return users;
 	}
 
 	/**
-	 * Takes in a user. 
-	 * Returns the userDataService update method with user as the parameter. 
+	 * Take in a user. 
+	 * Update credentials for user with credentialsDataService. 
+	 * If not updated, return 0. 
+	 * Return the userDataService update method using user. 
 	 * 
-	 * @param user	user to edit
+	 * @param user	user to update
 	 * @return int	result
 	 */
 	public int editUser(User user) {
+		
+		if (credentialsDataService.update(user.getCredentials()) != 1) {
+			return 0;
+		}
+		
 		return userDataService.update(user);
 	}
 
 	/**
-	 * Takes in a user. 
-	 * Returns the userDataService delete method with user as the parameter. 
+	 * Take in a user. 
+	 * Delete credentials for user with credentialsDataService. 
+	 * If not deleted, return 0. 
+	 * Return the userDataService delete method using user. 
 	 * 
-	 * @param user	user to remove
+	 * @param user	user to update
 	 * @return int	result
 	 */
 	public int removeUser(User user) {
+		
+		if (credentialsDataService.delete(user.getCredentials()) != 1) {
+			return 0;
+		}
+		
 		return userDataService.delete(user);
 	}
